Added double and complex for basic memory movement and scaling operations.

+++ src/dswap.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-void sswap(size_t N, float* X, int INCX, float* Y, int INCY){
+void sswap(size_t N, double* X, int INCX, double* Y, int INCY){
     int d = 0, ix = 0, iy = 0, tx = 0;
     for(d = 0; d < N; d++){
         tx = X[ix];
